+++ terraform/product/main.tf
@@ -18,18 +18,20 @@ module "haproxy" {
   config      = var.haproxy.config
 
   use_hacluster            = local.use_hacluster
+  hacluster_app_name       = local.use_hacluster ? var.hacluster.app_name : null
   hacluster_charm_channel  = local.use_hacluster ? var.hacluster.channel : null
   hacluster_charm_revision = local.use_hacluster ? var.hacluster.revision : null
   hacluster_config         = local.use_hacluster ? var.hacluster.config : {}
 
   use_keepalived            = local.use_keepalived
+  keepalived_app_name       = local.use_keepalived ? var.keepalived.app_name : null
   keepalived_charm_channel  = local.use_keepalived ? var.keepalived.channel : null
   keepalived_charm_revision = local.use_keepalived ? var.keepalived.revision : null
   keepalived_config         = local.use_keepalived ? var.keepalived.config : {}
 }
 
 resource "juju_application" "grafana_agent" {
-  name       = "grafana-agent"
+  name       = var.grafana_agent.app_name
   model_uuid = var.model_uuid
   units      = 1